LayoutTests/js/Promise-then-without-callbacks-in-workers-expected.txt

[Worker] Test Promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: resources/Promise-then-without-callbacks-in-workers.js
PASS [Worker] result is "hello"
PASS successfullyParsed is true

TEST COMPLETE
